pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "aichatcli"
version = "0.1.0"
description = "Fast multi-provider AI chatbot CLI"
authors = [{name = "GMfarhan", email = "your.email@example.com"}]
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
dependencies = [
    "python-dotenv>=1.0.0",
    "rich>=13.0.0", 
    "prompt-toolkit>=3.0.0"
]

[project.optional-dependencies]
groq = ["groq>=0.4.0"]
openai = ["openai>=1.0.0"]
anthropic = ["anthropic>=0.3.0"]
all = [
    "groq>=0.4.0",
    "openai>=1.0.0", 
    "anthropic>=0.3.0"
]

[project.scripts]
aichat = "AIchatCLI.cli:main"

[project.urls]
Homepage = "https://github.com/GMFarhan/hackathon-AI-chatbot"
Repository = "https://github.com/GMFarhan/hackathon-AI-chatbot"
